requirements.txt
Django==4.2.8
django-cors-headers==4.3.1
django-timezone-field==6.1.0
django-tinymce==3.6.1
djangorestframework==3.14.0
loguru==0.7.2
pytz==2023.3.post1
django-environ==0.11.2
gunicorn==21.2.0